+++ torii_boards/xilinx/artix7/alchitry_au.py
@@ -95,5 +95,5 @@ class AlchitryAuPlatform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	AlchitryAuPlatform().build(Blinky(), do_program = True)