+++ torii_boards/xilinx/artix7/nexys4ddr.py
@@ -201,5 +201,5 @@ class Nexys4DDRPlatform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	Nexys4DDRPlatform().build(Blinky(), do_program = True)