+++ torii_boards/xilinx/kintex7/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .genesys2 import Genesys2Platform
+from .kc705    import KC705Platform
+
+__all__ = (
+	'Genesys2Platform',
+	'KC705Platform',
+)