+++ torii_boards/xilinx/kintex7/genesys2.py
@@ -417,5 +417,5 @@ class Genesys2Platform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	Genesys2Platform().build(Blinky(), do_program = True)