+++ torii_boards/xilinx/kintex_us/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .kcu105 import KCU105Platform
+
+__all__ = (
+	'KCU105Platform',
+)