+++ torii_boards/xilinx/kintex_us/kcu105.py
@@ -41,5 +41,5 @@ class KCU105Platform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	KCU105Platform().build(Blinky(), do_program = True)