+++ torii_boards/xilinx/spartan3/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .mercury import MercuryPlatform
+
+__all__ = (
+	'MercuryPlatform',
+)