+++ torii_boards/xilinx/spartan6/__init__.py
@@ -1,0 +1,11 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .atlys        import AtlysPlatform
+from .numato_mimas import NumatoMimasPlatform
+from .sk_xc6slx9   import SK_XC6SLX9Platform
+
+__all__ = (
+	'AtlysPlatform',
+	'NumatoMimasPlatform',
+	'SK_XC6SLX9Platform',
+)