+++ torii_boards/xilinx/spartan6/numato_mimas.py
@@ -53,5 +53,5 @@ class NumatoMimasPlatform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	NumatoMimasPlatform().build(Blinky())