+++ torii_boards/xilinx/spartan7/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .arty_s7 import ArtyS7_25Platform, ArtyS7_50Platform
+from .cmod_s7 import CmodS7_Platform
+
+__all__ = (
+	'ArtyS7_25Platform',
+	'ArtyS7_50Platform',
+	'CmodS7_Platform',
+)