+++ torii_boards/xilinx/spartan7/cmod_s7.py
@@ -86,5 +86,5 @@ class CmodS7_Platform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	CmodS7_Platform().build(Blinky(), do_program = True)