+++ torii_boards/xilinx/zynq7/arty_z7.py
@@ -192,5 +192,5 @@ class ArtyZ720Platform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	ArtyZ720Platform().build(Blinky(), do_program = True)