+++ torii_boards/xilinx/zynq7/ebaz4205.py
@@ -47,5 +47,5 @@ class EBAZ4205Platform(XilinxPlatform):
 
 
 if __name__ == '__main__':
-	from ..test.blinky import Blinky
+	from ...test.blinky import Blinky
 	EBAZ4205Platform().build(Blinky(), do_program = True)